Rename LocationForegroundService.kt to GateGeofenceService.kt Add geofence constants Update .gitignore
